feat: 使用swc作为代码解析器
