+++ src/core/praser/swcParser.ts
@@ -60,28 +60,3 @@ type CollectFnType = (path:string) => void
          return m;
      }
  }
-
- /**
-  * 利用swc换转代码期间借用visitor实例访问ast获取模块路径
-  * @param code 
-  * @returns 
-  * @author chris lee
-  * @Time 2022/01/06
-  */
-export const collectImportNodes = async (code?:string):Promise<string[]> => {
-    if(!code) return [];
-    const result:string[] = [];
-    const collectPath = (path:string) => {
-        result.push(path);
-    };
-    const prasePlugin = new SWCVisitor(collectPath);
-    const output = await transform(code, {
-        plugin: (m) => prasePlugin.visitProgram(m),
-        jsc:{
-            parser: {
-                syntax: 'typescript',
-            }
-        }
-    });
-    return result;
-};
